axiom can now be tagged for better proof specification in cloning.
Axioms can now be tagged with a set of symbols. The syntax is:

  `axiom [tag1 tag2...]? nosmt? name : spec.`

In theory cloning, catch-all proof specification can be also tagged.
The syntax is:

  `clone ... proof T.* [tag1 tag2...]?`

The proof specification is then only applied to axioms whose tags
have a non-empty intersection with the proof specification tags.
(If no tags are provided, the old behavior is obtained - i.e. the
proof specification is applied unconditionally).

It is now allowed to give multiple starred proof-specifications. If so,
the first proof-specification matching is applied.