cloning: regexp based renamings
Work for types, operators, predicates, modules and module types.

The syntax is:

 - `clone ... rename [class1, ...] "regexp" to "subst" ...`
 - `clone ... rename "regexp" to "subst" ...`

where `class_i` is in `type`, `op`, `pred`, `module` and `module type`.

The `rename` clause applies the regexp to all the object names of the
cloned theories, whose class matches one of the given classes (or
regardless of the object class if no classes are given).
If the regexp matches, the new object name is given by the substitution
pattern. Once a rename clause matches, no more renaming is performed
for the object (i.e. the mechanism as a first match semantics).

See the PCRE documentation for help about the regexp and substitution
syntax. Note that the $0 substitution pattern is forbidden.