Crude detection of identity rewriting.
The mechanism is to forbid non-keyed matching in
`rewrite` when LHS && RHS of the rewriting equation
are convertible. Currently, the whole `rewrite` is rejected
if the matching succeeds but the result would result in a
identity rewriting. A better mechanism should embed
the identity detection rewriting mechanism into the form
matching one.